+++ src/main/java/bg/tu_varna/sit/b1/f22621620/source/equipment/Spell.java
@@ -85,7 +85,7 @@ public enum Spell {
         return stringBuilder.append("Spell: ")
                 .append(spell)
                 .append("; Damage: ")
-                .append(damage * 100)
+                .append((int) (damage * 100))
                 .toString();
     }
