+++ src/main/java/bg/tu_varna/sit/b1/f22621620/source/equipment/Weapon.java
@@ -45,7 +45,7 @@ public enum Weapon {
         return stringBuilder.append("Weapon: ")
                 .append(weapon)
                 .append("; Damage: ")
-                .append(damage * 100)
+                .append((int) (damage * 100))
                 .toString();
     }
